// src/app/shared/models/h-l-price/h-l-price.model.ts
export interface HLPriceResponesModel {
  _id: string;
  created: string;
  high99bid1: number;
  high99ask1: number;
  high96bid1: number;
  high96ask1: number;
  low99bid1: number;
  low99ask1: number;
  low96bid1: number;
  low96ask1: number;
  high99bid2: number;
  high99ask2: number;
  high96bid2: number;
  high96ask2: number;
  low99bid2: number;
  low99ask2: number;
  low96bid2: number;
  low96ask2: number;
  high99bid3: number;
  high99ask3: number;
  high96bid3: number;
  high96ask3: number;
  low99bid3: number;
  low99ask3: number;
  low96bid3: number;
  low96ask3: number;
  high99bid4: number;
  high99ask4: number;
  high96bid4: number;
  high96ask4: number;
  low99bid4: number;
  low99ask4: number;
  low96bid4: number;
  low96ask4: number;
  high96askCB: number;
  high96askCBDate: string;
  high96bidCB: number;
  high96bidCBDate: string;
  high99askCB: number;
  high99askCBDate: string;
  high99bidCB: number;
  high99bidCBDate: string;
  low96askCB: number;
  low96askCBDate: string;
  low96bidCB: number;
  low96bidCBDate: string;
  low99askCB: number;
  low99askCBDate: string;
  low99bidCB: number;
  low99bidCBDate: string;
  high96ask1Date: string;
  high96ask2Date: string;
  high96ask3Date: string;
  high96ask4Date: string;
  high96bid1Date: string;
  high96bid2Date: string;
  high96bid3Date: string;
  high96bid4Date: string;
  high99ask1Date: string;
  high99ask2Date: string;
  high99ask3Date: string;
  high99ask4Date: string;
  high99bid1Date: string;
  high99bid2Date: string;
  high99bid3Date: string;
  high99bid4Date: string;
  low96ask1Date: string;
  low96ask2Date: string;
  low96ask3Date: string;
  low96ask4Date: string;
  low96bid1Date: string;
  low96bid2Date: string;
  low96bid3Date: string;
  low96bid4Date: string;
  low99ask1Date: string;
  low99ask2Date: string;
  low99ask3Date: string;
  low99ask4Date: string;
  low99bid1Date: string;
  low99bid2Date: string;
  low99bid3Date: string;
  low99bid4Date: string;
}